Move ViewModel out of Adapter
